+++ src/context/ThemeContextWrapper.tsx
@@ -13,11 +13,15 @@ import type { ChildrenProp } from "../type/propTypes";
 //But Vite’s React plugin has a known quirk: a file exporting both component + hook + imports that aren’t components may still trigger the warning.
 // ? That is why const useTheme =() => {return useContext(ThemeContext)}; is moved to a separate file useTheme.js
 
-export const ThemeContextProvider = ({children}:ChildrenProp) => {
-    const [currentTheme, setCurrentTheme] = useState<Theme>(lightTheme);
+export const ThemeContextProvider = ({
+    children,
+    initialTheme = lightTheme,
+    secondTheme = darkTheme
+}:ChildrenProp & {initialTheme?:Theme, secondTheme?:Theme}) => {
+    const [currentTheme, setCurrentTheme] = useState<Theme>(initialTheme);
 
     const toggleTheme = () => {
-        setCurrentTheme((prevTheme:Theme) => (prevTheme === lightTheme ? darkTheme : lightTheme));
+        setCurrentTheme((prevTheme:Theme) => (prevTheme === initialTheme ? secondTheme : initialTheme));
     };
 
     return(