+++ CHANGELOG.md
@@ -28,3 +28,12 @@
 ### Added
 - Added labelAndInputContainerClass as conditional prop in DynamicFormProps
 - Added className FieldsetWrapper for distinct selector when styling using styled-components
+
+### [1.1.5] - 2025-12-03
+### Summary of Changes
+### Removed
+- delete logics for DynamicForm
+
+### Added
+- Changed edit logics to reset logics for DynamicForm
+- initialTheme and secondTheme props in ThemecontextProvider to allow override of theme